curr update: change password

+++ app/Http/Controllers/Master/UserController.php
@@ -4,16 +4,20 @@ namespace App\Http\Controllers\Master;
 
 use Carbon\Carbon;
 use App\Models\User;
-use Illuminate\Http\Request;
 use App\Models\Master\Branch;
 use App\Models\Master\Employee;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PasswordExpiredRequest;
+use App\Mail\SendAccount;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Storage;
 use RealRashid\SweetAlert\Facades\Alert;
 use Yajra\DataTables\Facades\DataTables;
-use Illuminate\Support\Facades\Validator;
-use App\Http\Requests\PasswordExpiredRequest;
 
 class UserController extends Controller
 {
@@ -33,9 +37,9 @@ class UserController extends Controller
                     ->addColumn('action', function($row){
 
                         $btn = '
-                        <form action="users/'.$row->id.'" method="POST" class="text-center">
+                        <form action="user/'.$row->empl_id.'" method="POST" class="text-center">
 
-                                <a class="btn btn-icon btn-bg-light btn-active-color-primary btn-md me-1" href="users/'.$row->id.'" >
+                                <a class="btn btn-icon btn-bg-light btn-active-color-primary btn-md me-1" href="user/'.$row->id.'" >
                                     <!--begin::Svg Icon | path: assets/media/icons/duotune/general/gen004.svg-->
                                     <span class="svg-icon svg-icon-muted svg-icon-2"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
                                     <path d="M21.7 18.9L18.6 15.8C17.9 16.9 16.9 17.9 15.8 18.6L18.9 21.7C19.3 22.1 19.9 22.1 20.3 21.7L21.7 20.3C22.1 19.9 22.1 19.3 21.7 18.9Z" fill="black"/>
@@ -44,7 +48,7 @@ class UserController extends Controller
                                     <!--end::Svg Icon-->
                                 </a>
 
-                                <a class="btn btn-icon btn-bg-light btn-active-color-warning btn-md me-1" href="users/'.$row->id.'/edit" >
+                                <a class="btn btn-icon btn-bg-light btn-active-color-warning btn-md me-1" href="user/'.$row->id.'/edit" >
                                     <!--begin::Svg Icon | path: assets/media/icons/duotune/general/gen055.svg-->
                                     <span class="svg-icon svg-icon-muted svg-icon-2"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
                                     <path opacity="0.3" fill-rule="evenodd" clip-rule="evenodd" d="M2 4.63158C2 3.1782 3.1782 2 4.63158 2H13.47C14.0155 2 14.278 2.66919 13.8778 3.04006L12.4556 4.35821C11.9009 4.87228 11.1726 5.15789 10.4163 5.15789H7.1579C6.05333 5.15789 5.15789 6.05333 5.15789 7.1579V16.8421C5.15789 17.9467 6.05333 18.8421 7.1579 18.8421H16.8421C17.9467 18.8421 18.8421 17.9467 18.8421 16.8421V13.7518C18.8421 12.927 19.1817 12.1387 19.7809 11.572L20.9878 10.4308C21.3703 10.0691 22 10.3403 22 10.8668V19.3684C22 20.8218 20.8218 22 19.3684 22H4.63158C3.1782 22 2 20.8218 2 19.3684V4.63158Z" fill="black"/>
@@ -93,32 +97,43 @@ class UserController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'empl_id' => 'required|string',
-            'empl_branch' => 'required|string',
             'max_session' => 'required|string',
             'expired_pwd' => 'required|string',
             'name' => 'required|string',
-            'email' => 'required|email',
-            'password' => 'required|string',
+            'username' => 'required|string|unique:users,username',
         ]);
 
         if ($validator->fails()) {
             Alert::toast('Ups, Terjadi Sesuatu yang Salah!', 'error');
             return redirect()->back()->withErrors($validator)->withInput();
         } else {
-            User::create([
+            $password = $request->role . $request->empl_id;
+            $empl_id = str_replace('-', '', $request->empl_id);
+
+            $new_data = User::create([
                 'coy_id' => Auth::user()->coy_id,
-                'empl_id' => $request->empl_id,
+                'empl_id' => $empl_id,
                 'empl_branch' => $request->empl_branch,
                 'max_session' => $request->max_session,
                 'expired_pwd' => Carbon::parse($request->empl_eff_date)->format('Y-m-d'),
                 'name' => $request->name,
-                'email' => $request->email,
-                'password' => $request->password,
+                'username' => $request->username,
+                'password' => $password,
                 'created_by' => Auth::user()->name,
                 'updated_by' => Auth::user()->name,
-            ]);
+            ])->addRole($request->role);
+
+            if ($request->file('copy_ktp_name')) {
+                $request->file('copy_ktp_name')->move(storage_path('app/public/user/ktp'), 'ktp-' . Carbon::now()->format('Y-m-d') . '.jpg');
+                User::where('id', $new_data->id)->update(['copy_ktp_name' => 'ktp-' . Carbon::now()->format('Y-m-d') . '.jpg']);
+            }
+            if ($request->file('copy_npwp_name')) {
+                $request->file('copy_npwp_name')->move(storage_path('app/public/user/npwp'), 'npwp-' . Carbon::now()->format('Y-m-d') . '.jpg');
+                User::where('id', $new_data->id)->update(['copy_npwp_name' => 'npwp-' . Carbon::now()->format('Y-m-d') . '.jpg']);
+            }
+
             Alert::toast('Data Berhasil Dibuat!', 'success');
-            return redirect()->route('users.index');
+            $this->sendMail($request->email, $request->username, $password);
         }
     }
 
@@ -141,7 +156,7 @@ class UserController extends Controller
      */
     public function edit($user)
     {
-        $data = User::where('id', $user)->first();
+        $data = User::join('role_user', 'users.empl_id', '=', 'role_user.user_id')->where('users.id', $user)->first();
         $employees = Employee::select('empl_id', 'empl_name')->where('coy_id', Auth::user()->coy_id)->get();
         $branches = Branch::select('branch_code', 'branch_name')->where('coy_id', Auth::user()->coy_id)->get();
 
@@ -158,10 +173,10 @@ class UserController extends Controller
     public function update(Request $request, $user)
     {
         $validator = Validator::make($request->all(), [
-            'empl_id' => 'required|string',
-            'empl_branch' => 'required|string',
+            'max_session' => 'required|string',
+            'expired_pwd' => 'required|string',
             'name' => 'required|string',
-            'email' => 'required|email',
+            'username' => 'required|string|unique:users,username',
         ]);
 
         if ($validator->fails()) {
@@ -169,20 +184,31 @@ class UserController extends Controller
             return redirect()->back()->withErrors($validator)->withInput();
         } else {
             User::where('id', $user)->update([
-                'coy_id' => Auth::user()->coy_id,
-                'empl_id' => $request->empl_id,
                 'empl_branch' => $request->empl_branch,
                 'max_session' => $request->max_session,
                 'expired_pwd' => Carbon::parse($request->empl_eff_date)->format('Y-m-d'),
                 'name' => $request->name,
-                'email' => $request->email,
+                'username' => $request->username,
                 'updated_by' => Auth::user()->name,
             ]);
+
             if ($request->password) {
                 User::where('id', $user)->update([
                     'password' => Hash::make($request->password),
                 ]);
             }
+
+            if ($request->file('copy_ktp_name')) {
+                Storage::delete('public/user/ktp/' . $request->old_copy_ktp_name);
+                $request->file('copy_ktp_name')->move(storage_path('app/public/user/ktp'), 'ktp-' . Carbon::now()->format('Y-m-d') . '.jpg');
+                User::where('id', $user)->update(['copy_ktp_name' => 'ktp-' . Carbon::now()->format('Y-m-d') . '.jpg']);
+            }
+            if ($request->file('copy_npwp_name')) {
+                Storage::delete('public/user/npwp/' . $request->old_copy_npwp_name);
+                $request->file('copy_npwp_name')->move(storage_path('app/public/user/npwp'), 'npwp-' . Carbon::now()->format('Y-m-d') . '.jpg');
+                User::where('id', $user)->update(['copy_npwp_name' => 'npwp-' . Carbon::now()->format('Y-m-d') . '.jpg']);
+            }
+
             Alert::toast('Data Berhasil Diperbarui!', 'success');
             return redirect()->route('users.index');
         }
@@ -191,9 +217,10 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    public function destroy($user)
     {
-        $user->delete();
+        User::where('empl_id', $user)->delete();
+        DB::table('role_user')->where('user_id', $user)->delete();
 
         Alert::toast('Status Data Berhasil Dihapus!', 'success');
         return redirect()->route('users.index');
@@ -216,5 +243,11 @@ class UserController extends Controller
             'expired_pwd' => Carbon::now()->toDateTimeString()
         ]);
         return redirect()->back()->with(['status' => 'Password changed successfully']);
+    }
+
+    function sendMail($email, $username, $password) {
+        Mail::to($email)->send(new SendAccount($username, $password));
+
+        return redirect()->route('users.index');
     }
 }
